requirements.txt
numpy==1.26.4
matplotlib==3.10.6
scikit-image==0.25.2
filterpy==1.4.5
opencv-python==4.12.0.88
pillow==11.3.0
transformers==4.41.0
python-dotenv==1.1.1
python-telegram-bot==22.4
ultralytics==8.3.200
sympy==1.14.0
tabulate==0.9.0
pandas==2.3.2
python-telegram-bot==22.4
python-dotenv
